refactor: manage clients exclusively via Supabase
