+++ app/clients/page.tsx
@@ -4,7 +4,7 @@ import { motion } from 'framer-motion'
 import AddClientModal from '../../components/AddClientModal'
 import ClientCard from '../../components/ClientCard'
 import Toast from '../../components/Toast'
-import { useClients, Client } from '@/lib/providers/ClientsProvider'
+import { useClients, Client } from '@/components/ClientsProvider'
 import { Plus, Search } from 'lucide-react'
 import { Input } from '../../components/ui/input'
 import { Button } from '../../components/ui/button'
